festival-api 패키지 업데이트
- Apis 스케쥴러 호출시 발생하는 에러 해결(Json 응답 메시지 첫줄이 xml 코드였음)
- 기존 Kopis 스케쥴러의 xml->json 형식 수작업 변환을 jackson라이브러리 도입으로 간결화

group-chatRoom 패키지 업데이트
- 모임채팅방 상세 조회(참여자만 가능)
- 모임채팅방 수정/삭제(방장만 가능)
- 임시 패키지 auth 도입(AuService)
검토 후 member.auth.service.AuthService에 통합 필요

festival-comment 패키지 업데이트
- 댓글 수정/삭제(작성자만 가능)

NotProd 업데이트
- Comment, chatRoom 테스트 데이터 주석 처리
- 로그인 서비스 구현 필요

build.gradle.kts에 jackson 의존성 추가

+++ src/main/java/com/ll/hfback/domain/group/chatRoom/serviceImpl/ChatRoomServiceImpl.java
@@ -1,5 +1,6 @@
 package com.ll.hfback.domain.group.chatRoom.serviceImpl;
 
+import com.ll.hfback.domain.group.chatRoom.auth.AuService;
 import com.ll.hfback.domain.group.chatRoom.dto.ChatRoomDto;
 import com.ll.hfback.domain.group.chatRoom.dto.DetailChatRoomDto;
 import com.ll.hfback.domain.group.chatRoom.entity.ChatRoom;
@@ -24,6 +25,7 @@ import java.util.stream.Collectors;
 public class ChatRoomServiceImpl implements ChatRoomService {
     private final ChatRoomRepository chatRoomRepository;
     private final MemberRepository memberRepository;
+    private final AuService auService;
 
     // 해당 게시글의 모든 모임 조회
     @Override
@@ -38,8 +40,13 @@ public class ChatRoomServiceImpl implements ChatRoomService {
     // 해당 게시글의 모임 상세 조회
     @Override
     @Transactional(readOnly = true)
-    public Optional<DetailChatRoomDto> searchById(Long id, String memberId) {
+    public Optional<DetailChatRoomDto> searchById(Long id) {
         Optional<ChatRoom> chatRoom = chatRoomRepository.findById(id);
+
+        // 사용자 검증 - 현재 로그인한 사용자의 ID를 가져오는 메서드
+        Long currentUserId = auService.getCurrentUserId();
+        String memberId = String.valueOf(currentUserId);
+
         return chatRoom
                 // joinMemberList에 memberId 포함 여부 확인
                 .filter(room -> room.getJoinMemberIdList().contains(memberId))
@@ -67,11 +74,16 @@ public class ChatRoomServiceImpl implements ChatRoomService {
     @Override
     @Transactional
     public DetailChatRoomDto convertToDetailChatRoomDto(ChatRoom chatRoom) {
-        // 기존 참여자 ID 리스트를 가져옴
+        // 기존 참여자/대기자 ID 리스트를 가져옴
         List<String> joinMemberIdList = chatRoom.getJoinMemberIdList();
+        List<String> waitingMemberIdList = chatRoom.getWaitingMemberIdList();
 
         // ID를 사용하여 각 Member 엔티티를 조회하고 닉네임 리스트를 생성
         List<String> joinMemberNicknames = joinMemberIdList.stream()
+                .map(id -> memberRepository.findById(Long.valueOf(id))
+                        .orElseThrow(() -> new IllegalArgumentException("해당 멤버가 존재하지 않습니다.")).getNickname())
+                .toList();
+        List<String> waitingMemberNicknames = waitingMemberIdList.stream()
                 .map(id -> memberRepository.findById(Long.valueOf(id))
                         .orElseThrow(() -> new IllegalArgumentException("해당 멤버가 존재하지 않습니다.")).getNickname())
                 .toList();
@@ -84,6 +96,7 @@ public class ChatRoomServiceImpl implements ChatRoomService {
                 chatRoom.getRoomTitle(),
                 chatRoom.getRoomContent(),
                 joinMemberNicknames,  // 닉네임 리스트 전달
+                waitingMemberNicknames,
                 chatRoom.getRoomMemberLimit(),
                 joinMemberNicknames.size()  // 참여자 수 전달
         );
@@ -101,16 +114,23 @@ public class ChatRoomServiceImpl implements ChatRoomService {
             throw new IllegalArgumentException("채팅방 참여 인원수는 10-100 사이의 값이어야 합니다.");
         }
 
+        // 사용자 검증 - 현재 로그인한 사용자의 ID를 가져오는 메서드
+        Long currentUserId = auService.getCurrentUserId();
+        String memberId = String.valueOf(currentUserId);
+
+        // 현재 로그인한 사용자의 member 객체를 가져오는 메서드
+        Member member = auService.getCurrentMember();
+
         // 작성자의 이름을 참여자 명단에 추가 및 변환
         List<String> joinMemberIdList = new ArrayList<>();
-        joinMemberIdList.add(String.valueOf(createChatRoomForm.getMember().getId())); // Member 객체의 이름을 추가
+        joinMemberIdList.add(memberId); // Member 객체의 ID를 추가
 
         // 비어있는 대기자 명단 생성 및 변환
         List<String> waitingMemberIdList = new ArrayList<>();
 
         ChatRoom chatRoom = ChatRoom.builder()
+                .member(member)
                 .festivalId(festivalId)
-                .member(createChatRoomForm.getMember())
                 .roomTitle(createChatRoomForm.getRoomTitle())
                 .roomContent(createChatRoomForm.getRoomContent())
                 .roomMemberLimit(createChatRoomForm.getRoomMemberLimit())
@@ -128,6 +148,13 @@ public class ChatRoomServiceImpl implements ChatRoomService {
     public void updateChatRoom(Long chatRoomId, @Valid UpdateChatRoomForm updateChatRoomForm) {
         ChatRoom chatRoom = chatRoomRepository.findById(chatRoomId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 모임이 존재하지 않습니다."));
+
+        // 사용자 검증 - 현재 로그인한 사용자의 ID를 가져오는 메서드
+        Long currentUserId = auService.getCurrentUserId();
+        if (!chatRoom.getMember().getId().equals(currentUserId)) {
+            throw new IllegalStateException("모임 채팅방 수정 권한이 없습니다.");
+        }
+
         chatRoom.setRoomTitle(updateChatRoomForm.getRoomTitle());
         chatRoom.setRoomContent(updateChatRoomForm.getRoomContent());
         chatRoom.setRoomMemberLimit(updateChatRoomForm.getRoomMemberLimit());
@@ -140,13 +167,24 @@ public class ChatRoomServiceImpl implements ChatRoomService {
     public void deleteChatRoom(Long chatRoomId) {
         ChatRoom chatRoom = chatRoomRepository.findById(chatRoomId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 모임이 존재하지 않습니다."));
+
+        // 사용자 검증 - 현재 로그인한 사용자의 ID를 가져오는 메서드
+        Long currentUserId = auService.getCurrentUserId();
+        if (!chatRoom.getMember().getId().equals(currentUserId)) {
+            throw new IllegalStateException("모임 채팅방 삭제 권한이 없습니다.");
+        }
+
         chatRoomRepository.delete(chatRoom);
     }
 
     // 해당 모임채팅방에 참여신청
     @Override
     @Transactional
-    public void applyChatRoom(Long chatRoomId, String memberId) {
+    public void applyChatRoom(Long chatRoomId) {
+        // 사용자 검증 - 현재 로그인한 사용자의 ID를 가져오는 메서드
+        Long currentUserId = auService.getCurrentUserId();
+        String memberId = String.valueOf(currentUserId);
+
         ChatRoom chatRoom = chatRoomRepository.findById(chatRoomId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 모임이 존재하지 않습니다."));
